+++ src/server/srv.rs
@@ -745,6 +745,8 @@ fn start_accept_thread(
                                                 msg = err.into_inner();
                                                 workers.swap_remove(next);
                                                 if workers.is_empty() {
+                                                    error!("No workers");
+                                                    thread::sleep(sleep);
                                                     break
                                                 } else if workers.len() <= next {
                                                     next = 0;